Never append {suffix} to config name
This may be dangerous if changing things like encryption password,
because an incorrect vfs cache might me reused.

In terms of data loss, this is _less_ dangerous than the default:
even if some unimportant detail of the configuration changes,
the existing vfs cache will be reused so pending changes will not be lost in limbo.